# Input data generation optimization