--- a/docs/helper_index.txt
+++ b/docs/helper_index.txt
@@ -63,6 +63,7 @@ Whenever a new generally-useful helper is added, it should be indexed here.
     Helpers for creating web elements like HTMLCanvasElement, OffscreenCanvas, etc.
 - {@link webgpu/util/shader}: Helpers for creating fragment shader based on intended output values, plainType, and componentCount.
 - {@link webgpu/util/texture/base}: General texture-related helpers.
+- {@link webgpu/util/texture/data_generation}: Helper for generating dummy texture data.
 - {@link webgpu/util/texture/layout}: Helpers for working with linear image data
     (like in copyBufferToTexture, copyTextureToBuffer, writeTexture).
 - {@link webgpu/util/texture/subresource}: Helpers for working with texture subresource ranges.