--- a/src/webgpu/api/operation/command_buffer/copyTextureToTexture.spec.ts
+++ b/src/webgpu/api/operation/command_buffer/copyTextureToTexture.spec.ts
@@ -20,38 +20,9 @@ import { makeBufferWithContents } from '../../../util/buffer.js';
 import { align } from '../../../util/math.js';
 import { physicalMipSize } from '../../../util/texture/base.js';
 import { kBytesPerRowAlignment, dataBytesForCopyOrFail } from '../../../util/texture/layout.js';
+import { DataArrayGenerator } from './data_generation.js';
 
-// Find the next square number that is greater than or equal to the given value.
-function nextSquare(value : number) {
-  const x = Math.ceil(Math.sqrt(value));
-  return x * x;
-}
-
-let initialData : Uint8Array;
-
-function GetInitialData(byteSize: number): Uint8Array {
-  const prevSize = initialData?.length ?? 0;
-
-  if (prevSize < byteSize) {
-    const newData = new Uint8Array(nextSquare(byteSize));
-
-    if (initialData) {
-        // Do a faster copy of any previous data that was generated.
-        newData.set(initialData);
-    }
-
-    for (let i = prevSize; i < newData.length; ++i) {
-        newData[i] = ((i ** 3 + i) % 251) + 1; // Have all initialData be non zero.
-    }
-    initialData = newData;
-  }
-
-  if (initialData.length == byteSize) {
-      return initialData;
-  }
-
-  return new Uint8Array(initialData.buffer, 0, byteSize);
-}
+const dataGenerator = new DataArrayGenerator();
 
 class F extends GPUTest {
   GetInitialDataPerMipLevel(
@@ -69,7 +40,7 @@ class F extends GPUTest {
       (textureSizeAtLevel.height / blockHeightInTexel);
 
     const byteSize = bytesPerBlock * blocksPerSubresource * textureSizeAtLevel.depthOrArrayLayers;
-    return GetInitialData(byteSize);
+    return dataGenerator.generateView(byteSize);
   }
 
   GetInitialStencilDataPerMipLevel(
@@ -84,7 +55,7 @@ class F extends GPUTest {
       textureSizeAtLevel.width *
       textureSizeAtLevel.height *
       textureSizeAtLevel.depthOrArrayLayers;
-    return GetInitialData(byteSize);
+    return dataGenerator.generateView(byteSize);
   }
 
   DoCopyTextureToTextureTest(