--- a/src/webgpu/api/operation/command_buffer/image_copy.spec.ts
+++ b/src/webgpu/api/operation/command_buffer/image_copy.spec.ts
@@ -63,6 +63,7 @@ import {
   kBytesPerRowAlignment,
   TextureCopyLayout,
 } from '../../../util/texture/layout.js';
+
 import { DataArrayGenerator } from './data_generation.js';
 
 interface TextureCopyViewWithRequiredOrigin {