Issue: #1757

<hr>

WIP change, still need to apply the same optimization to `generate_data` in `image_copy.spec.ts` and do a bit more cleanup.

Initial results show that this change takes running the full suite of `copyTextureToTexture` tests **from 71s to 20s on my machine.**