@toji 2022-08-24
Updated `image_copy.spec.ts` now as well. That one was a bit trickier, because one of it's uses of the generated data overlapped with another one, and as a result if the same underlying buffer was reused it's data returned by the first call would get corrupted by the second, but I addressed that by wrapping the generator in a class (shared by `copyTextureToTexture` and `image_copy`) and having a second instance for that use.

With this change in place both sets of tests now take **less than a quarter** of their previous times, and the data generation functions have practically disappeared from the profiler results.

Runtimes (different machine than the previous timing)

|                 |  copyTextureToTexture.spec.ts | image_copy.spec.ts |
|------------|---------------------------------|----------------------|
| Without Patch | 49s | 66s |
| With Patch| 12s | 25s |